only start pi-hole.service after network time synchronised
pi-hole.service on first boot needs a synchronised time so that the domain
list can be downloaded by http.
Add an After/Required dependency on time-sync.target.
This has downside that if there is no network connectivity then
pi-hole.service will not run and so pi-hole will not be configured.